pyproject.toml
[project]
name = "extractive-rag-qa-mlops-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.5.0",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "mlflow>=3.8.1",
    "numpy>=2.4.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "torch>=2",
    "transformers>=4.57.6",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pre-commit>=3.7.1",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=60"

[tool.coverage.run]
branch = true
source = ["src"]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
